app/requirements.txt
requests
streamlit
typing_extensions
matplotlib
numpy
pillow
tensorflow-cpu
